Create README.md
Mobilepay

Schoolproject Database for Mobil paymentsolution.

The database must include:

* User (name, address, cardInformation, registrationDate)
* Transactions (to, from, amount, date)
* A table, which can handle company payments
* A textfile for saving queries

First you need to use https://app.dbdesigner.net/ to draw a diagram of the tables that should be in the database.
Then you have to build the tables and fill in dummy data in them.
You must create queries for (a) user creation (b) examples of payments. Document your queries in a .txt, and with screenshots.
Create a Java application that can list all users and all transactions.
Expand the application so that you can make payments.